requirements.txt
tensorflow>=2.10.0
scikit-learn>=1.0.2
pandas>=1.3.5
numpy>=1.21.5
matplotlib>=3.5.1
seaborn>=0.11.2
apify-client>=1.0.0
kafka-python>=2.0.2
nltk>=3.7
xgboost>=1.6.1
pyyaml>=6.0
python-dotenv>=0.20.0
textblob>=0.17.1 